Changed the coding style guide to PSR-2 Added namespaces and now uses the PSR-4 autoloader
