+++ .gitignore
@@ -1,0 +1,9 @@
+vendor/
+build/
+bin/
+tests/index.php
+
+
+# Commit your application's lock file http://getcomposer.org/doc/01-basic-usage.md#composer-lock-the-lock-file
+# You may choose to ignore a library lock file http://getcomposer.org/doc/02-libraries.md#lock-file
+composer.lock
