+++ src/Exceptions/OdfException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Odtphp\Exceptions;
+
+class OdfException extends \Exception
+{
+}
