+++ src/Exceptions/PclZipProxyException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Odtphp\Exceptions;
+
+class PclZipProxyException extends \Exception
+{
+}
