+++ src/Exceptions/PhpZipProxyException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Odtphp\Exceptions;
+
+class PhpZipProxyException extends \Exception
+{
+}
