+++ src/Exceptions/SegmentException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Odtphp\Exceptions;
+
+class SegmentException extends \Exception
+{
+}
