+++ tests/simplecheck.php
